+++ effectful/tests/NonDetTests.hs
@@ -8,7 +8,9 @@ import Test.Tasty.HUnit
 import Effectful
 import Effectful.Dispatch.Dynamic
 import Effectful.Dispatch.Static
+import Effectful.Dispatch.Static.Primitive
 import Effectful.Internal.Env qualified as I
+import Effectful.Internal.Utils
 import Effectful.NonDet
 import Effectful.Reader.Static
 import Effectful.State.Dynamic
@@ -23,6 +25,7 @@ nonDetTests = testGroup "NonDet"
     [ testCaseSteps "local state"  $ test_state evalStateLocal  expectedLocalState
     , testCaseSteps "shared state" $ test_state evalStateShared expectedSharedState
     ]
+  , testCaseSteps "local mutable state" $ test_mutState expectedLocalState
   , testCaseSteps "different handlers are independent" test_independentHandlers
   , testCase "stale reference is detected after rollback" test_staleReferenceAfterRollback
   ]
@@ -67,6 +70,23 @@ test_state evalState expectedState step = runEff $ do
       modify @Int (+1)
       _<- (modify @Int (+2) >> empty) <|> (modify @Int (+4) >> empty)
       modify @Int (+8)
+
+test_mutState
+  :: (OnEmptyPolicy -> Int)
+  -> (String -> IO ())
+  -> IO ()
+test_mutState expectedState step = runEff $ do
+  runMutInt 0 . runNonDetBoth test $ \policy result -> do
+    liftIO . step $ show policy
+    s <- stateMutInt $ \s -> (s, 0)
+    U.assertEqual "result" Nothing (dropLeft result)
+    U.assertEqual "state" (expectedState policy) s
+  where
+    test :: (NonDet :> es, MutInt :> es) => Eff es ()
+    test = do
+      modifyMutInt (+1)
+      _<- (modifyMutInt (+2) >> empty) <|> (modifyMutInt (+4) >> empty)
+      modifyMutInt (+8)
 
 test_independentHandlers :: (String -> IO ()) -> Assertion
 test_independentHandlers step = runEff $ do
@@ -117,6 +137,35 @@ getStorageCapacity = unsafeEff $ \es -> do
 
 ----------------------------------------
 -- Helpers
+
+data MutInt :: Effect
+type instance DispatchOf MutInt = Static NoSideEffects
+newtype instance StaticRep MutInt = MutInt (IORef' Int)
+
+runMutInt :: Int -> Eff (MutInt : es) a -> Eff es a
+runMutInt s action = unsafeEff $ \es -> do
+  ref <- newIORef' s
+  inlineBracket
+    (consEnv (MutInt ref) relinkMutInt es)
+    unconsEnv
+    (unEff action)
+  where
+    relinkMutInt :: Relinker StaticRep MutInt
+    relinkMutInt = Relinker $ \_ (MutInt ref0) -> do
+      ref <- newIORef' =<< readIORef' ref0
+      pure $ MutInt ref
+
+stateMutInt :: MutInt :> es => (Int -> (a, Int)) -> Eff es a
+stateMutInt f = unsafeEff $ \es -> do
+  MutInt ref <- getEnv es
+  (r, s) <- f <$> readIORef' ref
+  writeIORef' ref s
+  pure r
+
+modifyMutInt :: MutInt :> es => (Int -> Int) -> Eff es ()
+modifyMutInt f = stateMutInt $ ((),) . f
+
+----
 
 data OuterEmpty :: Effect where
   OuterEmpty :: OuterEmpty m a